requirements.txt
dash
numpy
pandas
plotly
gunicorn
networkx
pyvis
webcolors
